Middleware Administrador

+++ routes/web.php
@@ -38,49 +38,49 @@ Route::get('/about', function () {
     return view('about');
 })->name('nosotros');
 
-Route::get('/roles', [RolController::class, 'index'])->name('roles.index');
+Route::get('/roles', [RolController::class, 'index'])->middleware('Isadmin:administrador')->name('roles.index');
 
-Route::get('/roles/create', [RolController::class, 'create'])->name('roles.create');
+Route::get('/roles/create', [RolController::class, 'create'])->middleware('Isadmin:administrador')->name('roles.create');
 
-Route::post('/roles/store', [RolController::class, 'store'])->name('roles.store');
+Route::post('/roles/store', [RolController::class, 'store'])->middleware('Isadmin:administrador')->name('roles.store');
 
-Route::get('/roles/edit/{rol}',[Rolcontroller::class, 'edit'])->name('roles.edit');
+Route::get('/roles/edit/{rol}',[Rolcontroller::class, 'edit'])->middleware('Isadmin:administrador')->name('roles.edit');
 
-Route::put('/roles/update/{rol}',[Rolcontroller::class, 'update'])->name('roles.update');
+Route::put('/roles/update/{rol}',[Rolcontroller::class, 'update'])->middleware('Isadmin:administrador')->name('roles.update');
 
-Route::get('/roles/mostrar/{rol}',[Rolcontroller::class, 'show'])->name('roles.mostrar');
+Route::get('/roles/mostrar/{rol}',[Rolcontroller::class, 'show'])->middleware('Isadmin:administrador')->name('roles.mostrar');
 
-Route::delete('roles/borrar/{rol}',[Rolcontroller::class, 'destroy'])->name('roles.borrar');
+Route::delete('roles/borrar/{rol}',[Rolcontroller::class, 'destroy'])->middleware('Isadmin:administrador')->name('roles.borrar');
 
-Route::get('/categorias',[CategoriaController::class, 'index'])->name('categorias.index');
+Route::get('/categorias',[CategoriaController::class, 'index'])->middleware('Isadmin:administrador')->name('categorias.index');
 
-Route::get('/categorias/crear',[CategoriaController::class, 'create'])->name('categorias.crear');
+Route::get('/categorias/crear',[CategoriaController::class, 'create'])->middleware('Isadmin:administrador')->name('categorias.crear');
 
-Route::post('/categorias/store', [CategoriaController::class, 'store'])->name('categorias.store');
+Route::post('/categorias/store', [CategoriaController::class, 'store'])->middleware('Isadmin:administrador')->name('categorias.store');
 
-Route::get('/categorias/edit/{categoria}',[CategoriaController::class, 'edit'])->name('categorias.edit');
+Route::get('/categorias/edit/{categoria}',[CategoriaController::class, 'edit'])->middleware('Isadmin:administrador')->name('categorias.edit');
 
-Route::put('/categorias/update/{categoria}',[CategoriaController::class, 'update'])->name('categorias.update');
+Route::put('/categorias/update/{categoria}',[CategoriaController::class, 'update'])->middleware('Isadmin:administrador')->name('categorias.update');
 
-Route::delete('categorias/delete/{categoria}',[CategoriaController::class, 'destroy'])->name('categorias.delete');
+Route::delete('categorias/delete/{categoria}',[CategoriaController::class, 'destroy'])->middleware('Isadmin:administrador')->name('categorias.delete');
 
 Route::get('/categorias/mostrar/{categoria}',[CategoriaController::class, 'show'])->name('categorias.mostrar');
 
-Route::get('/productos/index',[ProductoController::class,'index'])->name('productos.index');
+Route::get('/productos/index',[ProductoController::class,'index'])->middleware('Isadmin:administrador')->name('productos.index');
 
-Route::get('/productos/create',[ProductoController::class, 'create'])->name('productos.create');
+Route::get('/productos/create',[ProductoController::class, 'create'])->middleware('Isadmin:administrador')->name('productos.create');
 
-Route::post('/productos/store',[ProductoController::class, 'store'])->name('productos.store');
+Route::post('/productos/store',[ProductoController::class, 'store'])->middleware('Isadmin:administrador')->name('productos.store');
 
-Route::get('/productos/edit/{producto}',[ProductoController::class, 'edit'])->name('productos.edit');
+Route::get('/productos/edit/{producto}',[ProductoController::class, 'edit'])->middleware('Isadmin:administrador')->name('productos.edit');
 
-Route::put('/productos/update/{producto}',[ProductoController::class, 'update'])->name('productos.update');
+Route::put('/productos/update/{producto}',[ProductoController::class, 'update'])->middleware('Isadmin:administrador')->name('productos.update');
 
-Route::delete('/productos/delete/{producto}',[ProductoController::class, 'destroy'])->name('productos.delete');
+Route::delete('/productos/delete/{producto}',[ProductoController::class, 'destroy'])->middleware('Isadmin:administrador')->name('productos.delete');
 
 Route::get('/productos/mostrar/{producto}',[ProductoController::class, 'show'])->name('productos.mostrar');
 
-Route::resource('provincias', ProvinciaController::class);
+Route::resource('provincias', ProvinciaController::class)->middleware('Isadmin:administrador');
 
 Route::resource('users', UserController::class);
 
@@ -88,3 +88,8 @@ Route::get('login', [LoginController::class, 'showLoginForm'])->name('login');
 Route::post('login', [LoginController::class, 'login']);
 Route::post('logout', [LoginController::class, 'logout'])->name('logout');
 
+Route::middleware('Isadmin:admin')->group(function () {
+    Route::resource('users', UserController::class)->only([
+        'index',
+    ]);
+});
